rename cl_typing to typing.py
