+++ cleanlab/count.py
@@ -36,7 +36,7 @@ import numpy as np
 import warnings
 from typing import Tuple, Union, Optional
 
-from cleanlab.cl_typing import LabelLike
+from cleanlab.typing import LabelLike
 from cleanlab.internal.multilabel_utils import stack_complement, get_onehot_num_classes
 
 from cleanlab.internal.util import (
